+++ demos/projects/ST/b-l475e-iot01a/port/azure_iot_flash_platform.c
@@ -84,7 +84,7 @@ AzureIoTResult_t AzureIoTPlatform_Init( AzureADUImage_t * const pxAduImage )
     return xResult;
 }
 
-uint32_t AzureIoTPlatform_GetFlashBankSize()
+int64_t AzureIoTPlatform_GetFlashBankSize()
 {
     return FLASH_BANK_SIZE;
 }
